+++ js/views/CardView.js
@@ -1,10 +1,9 @@
 app.CardView = Backbone.View.extend({
-  tagName: 'div',
   template: _.template($('#card-template').html()),
 
   render: function(){
 
-    
+    $('.loading').html('<p>&nbsp;</p>');    
     $('#card_view').append(this.$el.html(this.template(this.model.toJSON())));
     app.cardgrid.init();
     
@@ -13,25 +12,28 @@ app.CardView = Backbone.View.extend({
   initialize: function(){
 
     var that = this;
-    this.model.fetch().done(function(){
-      if (that.model.get('dataCaptured') == false) {
+   
+    if (this.model.get('dataCaptured') == false) {
+      this.model.fetch().done(function(){
+        app.cardList.sortByField('lastmodified_ticks','descending');
         that.model.set('dataCaptured' , true);
-        that.render();
-      }
-      
-    });
-
-    //this.model.on('change', this.render, this);
-    this.model.on('destroy', this.remove, this); // remove: Convenience Backbone's function for removing the view from the DOM
-  },      
+        that.doRender();
+      });
+    } else {
+      this.doRender();
+    }
+  },
+  doRender: function() {
+    if (this.model.get('dataToBeRendered') == true) {
+      this.model.set('dataToBeRendered' , false);
+      this.render();
+    }
+  }, 
   events: {
 //    'dblclick label' : 'edit',
 //    'keypress .edit' : 'updateOnEnter',
 //    'blur .edit' : 'close',
  //   'click .toggle': 'toggleCompleted',
  //   'click .destroy': 'destroy'
-  },
-  destroy: function(){
-    this.model.destroy();
   }
 });